pyproject.toml
[tool.poetry]
name = "milheiroapp"
version = "0.1.0"
description = ""
authors = ["jlxmns <jlxmns@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-unfold = "^0.42.0"
psycopg2 = "^2.9.10"
django-registration = "^5.1.0"
django-allauth = "^65.2.0"
django-allauth-ui = "^1.5.1"
django-htmx = "^1.21.0"
django-crispy-forms = "^2.3"
crispy-tailwind = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
